Refactor code to use functions rather then classes.
