--- packages/mdx-embed/src/components/twitch/helper.tsx
@@ -1,18 +1,0 @@
-import {isEmptyString} from "../../utils";
-
-export default class Helper {
-
-    static constructTwitchURL(twitchId?: string, channel?: string, collection?: string, baseUrl?: string) {
-        const srcBlocks = [baseUrl];
-
-        if (twitchId !== undefined) {
-            srcBlocks.push(`&video=v${twitchId}`);
-        } else if (!isEmptyString(channel)) {
-            srcBlocks.push(`&channel=${channel}`);
-        } else if (!isEmptyString(collection)) {
-            srcBlocks.push(`&collection=${collection}`);
-        }
-
-        return srcBlocks.join('');
-    }
-}
